Lots of cleanup and refactoring. Most notably, make the
prefix-length property guint8.

Also fix keeping the route cache in sync when "modifying"
an existing route via `ip route change` or `ip route replace`.

(cherry picked from commit 514245b1d7f3ec63e5f5a4e1aa4727b38ceb1ba4)

+++ src/nm-ip4-config.h
@@ -79,7 +79,7 @@ void nm_ip4_config_merge (NMIP4Config *dst, const NMIP4Config *src, NMIPConfigMe
 void nm_ip4_config_subtract (NMIP4Config *dst, const NMIP4Config *src);
 void nm_ip4_config_intersect (NMIP4Config *dst, const NMIP4Config *src);
 gboolean nm_ip4_config_replace (NMIP4Config *dst, const NMIP4Config *src, gboolean *relevant_changes);
-gboolean nm_ip4_config_destination_is_direct (const NMIP4Config *config, guint32 dest, int plen);
+gboolean nm_ip4_config_destination_is_direct (const NMIP4Config *config, guint32 dest, guint8 plen);
 void nm_ip4_config_dump (const NMIP4Config *config, const char *detail);
 
 /* Gateways */
